refactor(ui): fast-refresh compliance batch 4 — 8 route files, 2-4 warnings each
Part of #7850. 18 of 246 react-refresh/only-export-components warnings.

Same TanStack Router pattern as batches 2/3, applied to files with
multiple locally-referenced components instead of just one: moved
each page's component(s) to a new `-<route>-page.tsx` sibling, keeping
only Route (plus, where the Route config needs it, its search schema)
in the route file.

Three needed the closure-over-Route -> standalone-hook swap (same
technique as batch 2/3): surfaces.tsx, sudo.index.tsx, and
validators.index.tsx each had a second component (SurfacesTable,
SudoTable, ValidatorsTable) also reading Route.useSearch()/fullPath.
sudo.index.tsx additionally re-exports `type SudoSearch` (type-only,
so it doesn't itself trip the rule) so the moved page component can
type its shared sudoQueryParams() helper without importing the schema
back.

design.primitives.tsx is the authoritative raw-values showcase page,
exempted from the no-restricted-syntax design-token rules in
eslint.config.ts by exact file path -- extended both ignore lists to
cover the new -design-primitives-page.tsx sibling the showcase content
moved into, so it keeps demonstrating raw values without new warnings.

validators-index-empty-action.test.ts source-scans validators.index.tsx
for the empty-state JSX (unreachable via rendering since the live feed
isn't empty) -- repointed its read target to the new sibling file that
JSX now lives in.

Zero behavior change; no external importers of any moved component
existed outside its own route file.

237 -> 219 warnings; 0 eslint errors. Live-verified all 8 routes
(/surfaces, /validators, /sudo, /design/primitives, /agents,
/delegate, /tools/ss58, /about) -- render correctly with live data,
no console errors.
